Fixing errors stopping build on Xcode 4 with SDK 10.6
Fixed build time errors appearing on Xcode 4. Builds to completion via
Xcode 4 application and xcodebuild on the command line.

Increased requirements to Mac OS X 10.6 or later. No specific changes
were made to restrict the code to 10.6. It should still be possible to
build with 10.4 support but this is not a goal.

Commented out introspection code in DKRuntimeHelper.m; will need to
investigate to see how this affects runtime behaviour.